IC : main screen, bottom nav implement
After logging in, I set the logo at the top of the main screen and implemented the navigation at the bottom.
The navigation at the bottom consists of Home, Search, Add Photos, Details, and User.
Each button is switched to each screen when clicked.
Currently, it is separated only by the color of the background and will be implemented one by one.