IC : add photo button implement, firebase storage integration
A button for adding photos was implemented.
Firebase's storage was linked to the app.
When you click the add photo button, you request access to the photo library and access it when you receive access.
When uploading photos, it was implemented to be stored in the Firebase storage.
In the future, we will implement the ability for photos to be posted to the app.