IC : database construction and upload and link
We built a database and linked it with photo uploads.
When uploading a photo, the contents are input and stored in a database.
Save timestamp, uid, userid ,favoriteaccount ,favorites, imgurl, explain.
Favorite elements will be used for the'Like' function.
We will identify users by userid.
Soon these will build a screen to be uploaded.