IC : profile setting, follow function implemetation
When selecting a profile picture, it is stored in a database and can be viewed on a user page.
When the other user clicks the follow button, the number of followers increases, and when the user clicks the follow button again, it is implemented.
The ability to view the posts of users you follow in the feed will be implemented in the future.